Refactor scroll behavior and add smooth scrolling
Refactor the global CSS file to add smooth scrolling behavior to the HTML element. This improves the user experience when navigating through the website.

Refactor package.json and add new dependencies

Added the "next-nprogress-bar" and "react-modal-video" dependencies to the project's package.json file. These dependencies will be used for implementing a progress bar and a modal video component.

Remove unused import statement in ApartmentFeatures component

Removed the unused import statement for the AspectRatio component in the ApartmentFeatures component. This improves code cleanliness and removes unnecessary dependencies.

Remove Footer component from the home page

Removed the Footer component from the home page. This was done to improve the layout and design of the page.

Add useWindowSize hook for responsive behavior

Added a new hook called useWindowSize to handle responsive behavior in the layout component. This hook returns the current window size and will be used for dynamically adjusting the layout based on the screen size.

Update review section with fade-in animation

Updated the review section with a fade-in animation using the framer-motion library. This adds a visually appealing effect when the section comes into view.

Update project section with fade-in animation

Updated the project section with a fade-in animation using the framer-motion library. This adds a visually appealing effect when the section comes into view.

Delete test page

Deleted the test page as it is no longer needed in the project.

Update layout component with page transition animation

Updated the layout component to include a page transition animation using the framer-motion library. This provides a smoother transition between pages.

Update CounterAnimation component with fade-in animation

Updated the CounterAnimation component with a fade-in animation using the framer-motion library. This adds a visually appealing effect when the component comes into view.

Update about page with CounterAnimation component

Updated the about page to use the CounterAnimation component for displaying the project count. This provides a more dynamic and engaging user experience.